Fix RCLC int parameter get (cherry-pick) (#272)
* Fix RCLC int parameter get


Update


Fix test


Update


Update examples


* updated: ubuntu:jammy ros-tooling/setup-ros@0.3.0


* updated documentation: build status for Rolling (#266)

* updated documentation: bloom release table for rolling ubuntu yammy for all packages.


* install phyhton dependendcy for ubuntu yammi transition - should be removed later.


* interactive installation


* removed installation of python3.9


* using docker_image: ubuntu:jammy


* using ubuntu-20.04


* setup-ros@0.3.0


* retry


* retry


* update text


* commented out unit tests ActionServerTest


* uncrustify


* removed dependencies.repos


* removed vcs-repo-file-url from ci.yml


* revert ActionServerTest


* uncrustify


* reverted changes in test_action_server.cpp


* Fix RCLC int parameter get


Update


Fix test


Update


Update examples

+++ rclc_parameter/test/rclc_parameter/test_parameter.cpp
@@ -151,7 +151,7 @@ TEST(Test, rclc_node_init_default) {
 
   // Get parameters
   bool param1;
-  int param2;
+  int64_t param2;
   double param3;
 
   rclc_parameter_get_bool(&param_server, "param1", &param1);
